requirements.txt
ttach==0.0.3
numpy==1.16.6
bayesian_optimization==1.2.0
torch==1.4.0
albumentations==0.4.5
catalyst==20.5
opencv_python==4.2.0.32
pretrainedmodels==0.7.4
segmentation_models_pytorch==0.1.0
imgaug==0.4.0
pandas==1.0.4
scipy==1.3.3
pytorch_toolbelt==0.3.2
Pillow==7.2.0
